+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Improved parser rules for ClickHouse.
 - Improved migration script generation for tables and indexes with options in MS SQL.
+- Added the IS JSON parser rule for PostgreSQL.
 
 ### Fixed